+++ src/lib/abstract/abstract.request.interceptor.ts
@@ -1,11 +1,12 @@
 import { Logger, NotFoundException } from '@nestjs/common';
 import { plainToClass } from 'class-transformer';
 import { validate } from 'class-validator';
+import { Request } from 'express';
 import _ from 'lodash';
 import { BaseEntity } from 'typeorm';
 
 import { CreateParamsDto } from '../dto/params.dto';
-import { Column, GROUP } from '../interface';
+import { Author, Column, CrudOptions, GROUP, Method } from '../interface';
 
 export abstract class RequestAbstractInterceptor {
     async checkParams(
@@ -30,5 +31,22 @@ export abstract class RequestAbstractInterceptor {
             throw exception;
         }
         return Object.assign({}, transformed);
+    }
+
+    getAuthor(
+        request: Request | Record<string, unknown>,
+        crudOptions: CrudOptions,
+        method: Exclude<Method, Method.READ_MANY | Method.READ_ONE | Method.SEARCH>,
+    ): Author | undefined {
+        if (!crudOptions.routes || !crudOptions.routes[method] || !crudOptions.routes[method]?.author) {
+            return;
+        }
+
+        return {
+            ...crudOptions.routes[method]!.author!,
+            value:
+                crudOptions.routes[method]!.author?.value ??
+                _.get(request, crudOptions.routes[method]!.author!.filter!, crudOptions.routes[method]!.author?.value),
+        };
     }
 }
