+++ src/lib/interceptor/upsert-request.interceptor.ts
@@ -8,7 +8,6 @@ import { Observable } from 'rxjs';
 import { RequestAbstractInterceptor } from '../abstract';
 import { Constants } from '../constants';
 import { CrudOptions, CrudUpsertRequest, FactoryOption, GROUP, Method } from '../interface';
-import { AuthorHelper } from '../provider';
 
 export function UpsertRequestInterceptor(crudOptions: CrudOptions, factoryOption: FactoryOption) {
     class MixinInterceptor extends RequestAbstractInterceptor implements NestInterceptor {
@@ -38,7 +37,7 @@ export function UpsertRequestInterceptor(crudOptions: CrudOptions, factoryOption
             const crudUpsertRequest: CrudUpsertRequest<typeof crudOptions.entity> = {
                 params,
                 body,
-                author: AuthorHelper.extract(req, crudOptions, Method.UPSERT),
+                author: this.getAuthor(req, crudOptions, Method.UPSERT),
             };
 
             req[Constants.CRUD_ROUTE_ARGS] = crudUpsertRequest;
