+++ src/lib/interface/author.interface.ts
@@ -1,0 +1,5 @@
+export interface Author {
+    property: string;
+    filter?: string;
+    value?: unknown;
+}
