+++ src/lib/provider/index.ts
@@ -1,3 +1,2 @@
 export * from './execution-context-host.mock';
 export * from './pagination.helper';
-export * from './author.helper';
